requirements.txt
alembic==1.12.0
annotated-types==0.6.0
anyio==4.0.0
asyncpg==0.28.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
dadata==21.10.1
greenlet==2.0.2
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-dotenv==1.0.0
python-telegram-bot==20.6
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.21
typing_extensions==4.8.0
